scripts/source-health/requirements.txt
# Source Health Check System Dependencies

# Web scraping and browser automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# JSON schema validation
jsonschema>=4.20.0

# CLI and output
rich>=13.7.0
click>=8.1.0

# AI providers (optional - for repair functionality)
google-generativeai>=0.3.0
openai>=1.6.0
anthropic>=0.8.0

# Utilities
pyyaml>=6.0.0
python-dateutil>=2.8.0
